Implement downloader and uploader; use them in getProfile endpoint for large jfr files
